+++ Ex-1/src/hmorita/abexercise/entity/Keyable.java
@@ -3,5 +3,5 @@ package hmorita.abexercise.entity;
 import java.util.List;
 
 public interface Keyable {
-    public int key();
+    public long key();
 }
